+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `ready-symbol-shorthands` used in some sub-packages to keep their
   name lengths in check.
+- `buffer-name` segment now uses the local map and help echo from
+  `modeline-buffer-identification`.
 
 ## [0.4.0]